refactor(spf): move select-tracks orchestrations to playback/behaviors/
Splits packages/spf/src/media/primitives/select-tracks.ts into pure logic
(stays in media/primitives/) and signal-based orchestrations
(selectVideoTrack, selectAudioTrack, selectTextTrack — moved to
playback/behaviors/select-tracks.ts). Drops core/signals dependency from
media/primitives/. Tests split correspondingly.

This is the second of three media-→-core boundary fixes.
