+++ packages/spf/src/media/primitives/select-tracks.ts
@@ -1,5 +1,3 @@
-import { effect } from '../../core/signals/effect';
-import { type Signal, update } from '../../core/signals/primitives';
 import { DEFAULT_QUALITY_CONFIG, selectQuality } from '../abr/quality-selection';
 import type { AudioSelectionSet, Presentation, TrackType, VideoSelectionSet } from '../types';
 import { SelectedTrackIdKeyByType } from '../utils/track-selection';
@@ -262,107 +260,4 @@ export function shouldSelectTrack<T extends TrackType>(
   config: TrackSelectionConfig<T>
 ): boolean {
   return !state[SelectedTrackIdKeyByType[config.type]];
-}
-
-// =============================================================================
-// Orchestrations
-// =============================================================================
-
-/**
- * Select video track orchestration.
- *
- * Selects video track when:
- * - Presentation exists
- * - No video track is selected yet
- *
- * Uses bandwidth-based quality selection algorithm.
- *
- * @example
- * const cleanup = selectVideoTrack(
- *   { state, owners, events },
- *   { initialBandwidth: 2_000_000 }
- * );
- */
-export function selectVideoTrack<S extends TrackSelectionState>(
-  { state }: { state: Signal<S> },
-  config: VideoSelectionConfig = { type: 'video' }
-): () => void {
-  return effect(() => {
-    const currentState = state.get();
-    if (!canSelectTrack(currentState, config) || !shouldSelectTrack(currentState, config)) return;
-
-    // Just to have basic functionality/POC, simply selecting the first track
-    const selectedTrackId = currentState.presentation?.selectionSets.find(({ type }) => type === config.type)
-      ?.switchingSets[0]?.tracks[0]?.id;
-
-    if (selectedTrackId) {
-      const selectedTrackKey = SelectedTrackIdKeyByType[config.type];
-      const patch: Partial<TrackSelectionState> = { [selectedTrackKey]: selectedTrackId };
-      update(state, patch);
-    }
-  });
-}
-
-/**
- * Select audio track orchestration.
- *
- * Selects audio track when:
- * - Presentation exists
- * - No audio track is selected yet
- *
- * Uses language and preference-based selection.
- *
- * @example
- * const cleanup = selectAudioTrack(
- *   { state, owners, events },
- *   { preferredAudioLanguage: 'en' }
- * );
- */
-export function selectAudioTrack<S extends TrackSelectionState>(
-  { state }: { state: Signal<S> },
-  config: AudioSelectionConfig = { type: 'audio' }
-): () => void {
-  return effect(() => {
-    const currentState = state.get();
-    if (!canSelectTrack(currentState, config) || !shouldSelectTrack(currentState, config)) return;
-
-    // Just to have basic functionality/POC, simply selecting the first track
-    const selectedTrackId = currentState.presentation?.selectionSets.find(({ type }) => type === 'audio')
-      ?.switchingSets[0]?.tracks[0]?.id;
-
-    if (selectedTrackId) {
-      const patch: Partial<TrackSelectionState> = { selectedAudioTrackId: selectedTrackId };
-      update(state, patch);
-    }
-  });
-}
-
-/**
- * Select text track orchestration.
- *
- * Selects text track when:
- * - Presentation exists
- * - No text track is selected yet
- *
- * Note: Currently does not auto-select (user opt-in).
- *
- * @example
- * const cleanup = selectTextTrack({ state, owners, events }, {});
- */
-export function selectTextTrack<S extends TrackSelectionState>(
-  { state }: { state: Signal<S> },
-  config: TextSelectionConfig = { type: 'text' }
-): () => void {
-  return effect(() => {
-    const currentState = state.get();
-    if (!canSelectTrack(currentState, config) || !shouldSelectTrack(currentState, config)) return;
-
-    // Text tracks are user opt-in - don't auto-select
-    const selectedTextTrackId = pickTextTrack(currentState.presentation!, config);
-
-    if (selectedTextTrackId) {
-      const patch: Partial<TrackSelectionState> = { selectedTextTrackId };
-      update(state, patch);
-    }
-  });
 }
